YLW-3181 chore: Renaming files and namespace. Used const in tests
